Refactor/cleanup new websocket handling code
(c) Copyright IBM Corp. 2016
